refactor(agent): deepen workspace agent metadata
